% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{fred_category}
\alias{fred_category}
\title{Fetch FRED category info}
\usage{
fred_category(category_id = 0, api_key)
}
\arguments{
\item{category_id}{Category ID (default 0 = root)}

\item{api_key}{FRED API key}
}
\value{
tibble: id, name, parent_id
}
\description{
Fetch FRED category info
}
