% Generated by build script
\name{fred_category_children}
\alias{fred_category_children}
\title{Fetch child categories}
\usage{
fred_category_children()
}
\arguments{
\item{category_id}{Parent category ID}
\item{api_key}{FRED API key}
}
\value{tibble: id, name, parent_id}
\description{Fetch child categories}
